requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
frozenlist==1.2.0
future==0.18.2
idna==3.3
line-bot-sdk==2.0.1
multidict==5.2.0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
transitions==0.8.10
typing-extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
